+++ phoenix-core-server/src/main/java/org/apache/phoenix/coprocessor/CompactionScanner.java
@@ -207,8 +207,9 @@ public class CompactionScanner implements InternalScanner {
     boolean guardEnabled = conf.getBoolean(QueryServices.REPLICATION_COMPACTION_GUARD_ENABLED,
       QueryServicesOptions.DEFAULT_REPLICATION_COMPACTION_GUARD_ENABLED);
     if (replayEnabled && guardEnabled) {
-      this.maxLookbackWindowStart = applyReplicationConsistencyGuard(this.maxLookbackWindowStart,
-        conf, tableName, columnFamilyName);
+      this.maxLookbackWindowStart =
+        ReplicationLogReplayService.applyReplicationConsistencyGuard(this.maxLookbackWindowStart,
+          conf, tableName, columnFamilyName);
     }
     ColumnFamilyDescriptor cfd = store.getColumnFamilyDescriptor();
     this.major = major && !forceMinorCompaction;
@@ -370,44 +371,6 @@ public class CompactionScanner implements InternalScanner {
       : maxLookbackMap.get(tableName + CompactionScanner.SEPARATOR + columnFamilyName);
   }
 
-  /**
-   * Applies the replication replay consistency point as a floor on maxLookbackWindowStart. On
-   * standby clusters, this prevents compaction from dropping delete markers that have timestamps
-   * newer than the consistency point.
-   */
-  @VisibleForTesting
-  static long applyReplicationConsistencyGuard(long currentMaxLookbackWindowStart,
-    Configuration conf, String tableName, String columnFamilyName) {
-    try {
-      long consistencyPoint = getConsistencyPointFromReplayService(conf);
-      return adjustMaxLookbackWindowStart(currentMaxLookbackWindowStart, consistencyPoint,
-        tableName, columnFamilyName);
-    } catch (Exception e) {
-      LOGGER
-        .warn("Replication guard enabled but consistency point unavailable for table={} store={}."
-          + " Retaining all delete markers.", tableName, columnFamilyName, e);
-      return 0L;
-    }
-  }
-
-  @VisibleForTesting
-  static long getConsistencyPointFromReplayService(Configuration conf) throws Exception {
-    ReplicationLogReplayService service = ReplicationLogReplayService.getInstance(conf);
-    return service.getConsistencyPoint();
-  }
-
-  @VisibleForTesting
-  static long adjustMaxLookbackWindowStart(long currentMaxLookbackWindowStart,
-    long consistencyPoint, String tableName, String columnFamilyName) {
-    long adjusted = Math.min(currentMaxLookbackWindowStart, consistencyPoint);
-    if (adjusted < currentMaxLookbackWindowStart) {
-      LOGGER.info(
-        "Replication guard: table={} store={} maxLookbackWindowStart adjusted from {} to {}"
-          + " (consistencyPoint={})",
-        tableName, columnFamilyName, currentMaxLookbackWindowStart, adjusted, consistencyPoint);
-    }
-    return adjusted;
-  }
 
   static class CellTimeComparator implements Comparator<Cell> {
     public static final CellTimeComparator COMPARATOR = new CellTimeComparator();
